bcache: Fix a sector counting race
With btree cmpxchg operations, the bch_add_sectors() call at the top of
bch_extent_insert_fixup() is for a key that isn't necessarily currently in the
btree; and whatever pointers we weren't adding/moving won't be pinned by open
buckets.

So it's possible to race with foreground writes/the allocator and have pointers
in that key that we think should be dirty, but actually aren't or might even be
stale.

Worse, we could have add_sectors() work as expected - but then have bkey_cmpxg()
fail, and between the add_sectors() call and the subtract_sectors() call in the
failure path have one of the pointers go stale - and then the two different
calls will disagree about what pointers are dirty and which aren't, leading to
badness.

Fix is to
 - not consider stale pointers to be replicas that weren't found, for general
   sanity, so that the interpretation of a key can't change from one moment to
   the next, and

 - have bch_extent_insert_fixup() bail out early if the first add_sectors()
   found a stale pointer, so that we aren't adding/subtracting sector counts
   from a bucket that's been reused (the bkey_cmpxchg() call will fail anyways).

Change-Id: I1a97c9a21d507c7db50b77a61f5774b5a7c2e001