dfs client: try to handle case where newer version is in older bset
Previously the code was missing a bkey_written() case. Attempting to
add this, while also cleaning up the existing logic a little bit.
We need more unit tests to check this case.

Also add an assertion to bch_bset_insert() to catch the case where
the existing bkey is out of bounds. I suspect we hit this in testing,
and the memmove() inappropriately moved data spanning multiple bsets.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>