bcache: Make some parameters sysfs-controllable and keybufs resizeable.
Make sectors_until_gc and reserve capacity sysfs-configurable.

Change keybuf so that it can be resized by sysfs and add code to tear
them down in suitable places.

Document and allow ref counts in keybuf_key to be -1 when freeing the
keys in the rb tree as they indicate an IO that was never started, as
opposed to 0 which indicates one that was finished.

Change-Id: Ia87c61cf14b7cf2ef890169324711b2500f46d47
Signed-off-by: Bill Rozas <gjr@datera.io>