bcache: Avoid memmove when dequeuing keylist entries.  Use pointers.
bch_keylist_pop_front used to memmove from back to front to dequeue
an element.  This leads to quadratic time if a keylist is long.

Add another pointer to the current front, and bump that instead.

Rename the fields of a keylist with changed semantics to catch all users.

Add a pointer to the end of the currently-allocated array.

Add a few asserts (BUG_ON).

Change-Id: I0b6a8ac2f76e63e7351d5b224b9fa5762975ae3b
Signed-off-by: Bill Rozas <gjr@datera.io>