bcache: Fix btree code for handle_existing_key_newer()
the btree code previously assumed that a single extent insertion required a
bounded amount of space. handle_exinting_key_newer() broke that assumption - we
now have to be able to partially insert an extent.

required lots of contortions, and it was pretty ugly, and testing this thing
scares me, but - this should bail out and let the rest be inserted later if we
run out of space.

Change-Id: Icd4cd4e3cea4d9366f82d846d10770b38eadb748