bcache: fix deadlock when setting new root
If the journal was full and the oldest node in the journal was the
old root, we would self-deadlock from trying to take a read lock on
the old root while already holding a write lock.

To get around this, push the write lock down into btree_set_root(),
and take it after obtaining a journal reservation.

Change-Id: I85aa5f63c8ebd4aaa9e566122585379ac89fc620