bcache: fix tiering_pred to not crash with funky pointers
If we have too many pointers on tier 0, we would try to copy the
extent and possibly hit the assertion checking that a key has on
more than BKEY_EXTENT_PTR_MAX (4) pointers.

This should not happen in normal operation, but will come up with
device removal data migration until other fixes are made to the
migration logic.

Change-Id: I414d823d42e8f7621ec9e08dd9447a7edeb938fe