bcache: Mark the moving_io structures for btree gc.
Add a write_pending list for moving ios whose writes have been issued
but not yet finished, so that we can still mark them.

In addition, actually mark the (replacement) keys in the pending
and write_pending lists.

Finally, reorder the calls to bch_data_move with the calls to dequeue
the key from the scan list, to remove a window where a key would not
be found in either.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>