bcache: Add a read-with-versions ioctl.
The read-with-versions ioctl picks up the versions in the initial
pass, before returning EIOCBQUEUED to the user.

As this can be incorrect if any of the constituent reads retries,
for now, if that happens, and the read was a versioned read, we
return EAGAIN for the whole ioctl.

Later, we'll fix the retries to avoid returning EAGAIN.

Change-Id: If3a9e7d497ee839790cf191e2d98e737ec3ced5c