aio: Convert to closure waitlist for aio ring buffer
Advantage of closure waitlists is that we don't have to muck with the task state
before we actually sleep; instead of prepare_to_wait() we do closure_wait(),
which like prepare_to_wait() adds an object to a waitlist but unlike
prepare_to_wait it's the closure that's doing the waiting, not the task.

This fixes the issue with doing copy_to_user() after modifying the task state.

Change-Id: Ifc75123d5bb620277d1e78dd5102e5d8bead1add