bcache: try to write multiple btree nodes when journal is full
When we are out of journal buckets: instead of only writing out
the oldest node, write out enough nodes to free up a journal bucket
on each journal device.

When the pin fifo is full: write out enough nodes to free up 1/16th
of the FIFO.

Finally, increase the journal.pin FIFO size to match the max journal
size of 512MB. In the future, this FIFO should be made resizable.

Also add some comments attempting to clarify how the journal code
works.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>