bcache: fix journal after variable size bucket changes
Now, we only use the minimum of the bucket sizes across cache
devices in tier 0. This is because the logic to handle variable
size buckets was broken in the case where we have some space in
one of the journal buckets, but not enough for a journal entry.

In this case, journal_reclaim() would enter an infinite loop,
because __bch_journal_res_get() would set c->journal.blocks_free
to zero to signal that there's not enough space for the requested
reservation, but then journal_reclaim() would just recompute
c->journal.blocks_free from ja->blocks_free, and loop again, etc.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>