bcache: Don't waste journal space when using different bucket sizes
This is readding the functionality Jacob added in 96bcb49472, but with the bug
fixed that Slava fixed by reverting the functionality in dca7372f4a.

Also count remaining space in the journal bucket(s) in sectors, instead of
blocks (actually changing the units this time!), and do a bit of other
refactoring.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>