bcache: fix deadlock when cache set goes read only if writes were throttled
When bch_cache_set_read_only() was added, we forgot to change
bch_wake_delayed_writes() to check for CACHE_SET_RO as well as
CACHE_SET_STOPPING. As a result, it would possibly not wake up
all writes. Since it was only called once, some writes could
remain on the throttle list forever and we would hang.

Change-Id: Ic51ea813bb57b04183b846244b85792ae02d9596