bcache: Don't allow 0 size extents that overlap with other extents
We never allowed a 0 size extent to overlap with another extent in the same bset
- but it could overlap with another extent in the same btree node due to extent
merging:

                   || <- zero size extent	bset 0
   |existing extent||new extent|		bset 1

post merge:
		   ||
   |merged extent              |

This means that if you're visiting keys in order ordered by the ending offset of
the key, they won't also be in order by starting offset - and
bch_insert_fixup_extent() needs to visit keys in order by starting offset, so
the btree node iter's comparison function has been forced to compute and compare
starting offsets.

This patch checks for and fixes up these 0 size extents, so this situation won't
occur and we'll be able to use a cheaper comparison function for the btree node
iterator.

Change-Id: I3cdfb3cc1e8d53daf1c55a2adbf1d535ea7a3f24