bcache: Hack around a deadlock in journal replay
Journal replay might need to set a new btree root, which requires writing a
journal entry - thus, if the journal is completely full and this happens when
replaying the first journal bucket's entries we're screwed.

Hack around this by not letting the journal fill up unless we're in journal
replay.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>