bcache: Fix a race in the async splits free path
bch_pending_btree_node_free_insert_done() -> bch_mark_pointers() was racing with
the start of gc, it was taking btree_node_pending_free_lock to avoid racing with
gc running past the GC_PHASE_PENDING_DELETE position but that was insufficient.
just hack around this with the big hammer.