bcache: rework alloc_failed()
Rework the way mark and sweep gc and the allocator threads are kicked:

Now, mark and sweep gc is only kicked from the allocator thread when the
allocator thread can't find free buckets, and when there are buckets with
saturated counters _and_ we have tried to subtract one allocator batch's worth
of sectors from saturated buckets since the last gc.

The allocator thread now explicitly waits on the number of available buckets,
and is kicked from buckets.c when we make buckets available, and garbage
collection when we reset the bucket counts at the end.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>