bcache: Move index lookup out from under write lock
This is a relatively minor performance improvement (perhaps a significant one
when creating snapshots, but that hasn't been tested) - we move the index lookup
within the leaf node (initializing the btree iterator) out from under the write
lock. This also means doing multiple insertions without reiniting the iterator.

The immediate need for this patch is that it's prep work for fixing
handle_existing_key_newer - the direction this work is pushing the code turned
out to be exactly what was required for that patch.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>